# Boost indexing performance by using parallel iterators