--- a/Cargo.toml
+++ b/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+jwalk = "0.8.1"
 serde = { version = "1.0.152", features = ["derive"] }
 serde_json = "1.0.91"
 sqlite = "0.30.3"