@ghassanachi 2023-03-04
I just tested the `Arc<Mutex<Model>>` solution and there was about a 15% regression in performance. I didn't profile it, but my guess would be that lock-contention is introducing more of a bottleneck and ends up being slower than the collect call. 

With that in mind, I'll move this PR from a Draft state.

Let me know if you want me to add symlink support.